GNS3 installer for Arch Linux
the script has been created to install the gns3 networking environment, this script installs all the packages necessary to have a "ready to use" configuration without problems due to the lack of some simulator dependencies. for further information, the author recommends looking at the Arch documentation. The author is not responsible for inadvertent damage to other people's computer systems